requirements.txt
scrapy>=2.11
psycopg2-binary